+++ mage/src/main/java/mil/nga/giat/mage/LandingActivity.java
@@ -262,13 +262,6 @@ public class LandingActivity extends AppCompatActivity implements NavigationView
             // notify location services that the permissions have changed.
             ((MAGE) getApplication()).getLocationService().onLocationPermissionsChanged();
         }
-
-        try {
-            Event event = EventHelper.getInstance(getApplicationContext()).getCurrentEvent();
-            setAvailableLayersIcons(LayerHelper.getInstance(getApplicationContext()).readByEvent(event, "GeoPackage"));
-        } catch (LayerException e) {
-            Log.e(LOG_NAME, "Error reading layers", e);
-        }
     }
 
     @Override
@@ -321,17 +314,6 @@ public class LandingActivity extends AppCompatActivity implements NavigationView
     private void setTitle() {
         Event event = EventHelper.getInstance(getApplicationContext()).getCurrentEvent();
         getSupportActionBar().setTitle(event.getName());
-    }
-
-    private void setAvailableLayersIcons(Collection<Layer> layers) {
-        Collection<Layer> available = Collections2.filter(layers, new Predicate<Layer>() {
-            @Override
-            public boolean apply(Layer layer) {
-                return !layer.isLoaded();
-            }
-        });
-
-        LandingActivity.this.findViewById(R.id.available_layer_downloads).setVisibility(available.isEmpty() ? View.GONE : View.VISIBLE);
     }
 
     private void showDisabledPermissionsDialog(String title, String message) {