+++ mage/src/main/java/mil/nga/giat/mage/observation/LocationEditActivity.java
@@ -3,11 +3,13 @@ package mil.nga.giat.mage.observation;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.res.Configuration;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.os.Vibrator;
+import android.preference.PreferenceManager;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v4.graphics.drawable.DrawableCompat;
@@ -163,6 +165,9 @@ public class LocationEditActivity extends AppCompatActivity implements TextWatch
     public void onMapReady(GoogleMap map) {
         this.map = map;
         map.setOnCameraIdleListener(this);
+
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
+        map.setMapType(preferences.getInt(getString(R.string.baseLayerKey), getResources().getInteger(R.integer.baseLayerDefaultValue)));
 
         int dayNightMode = getResources().getConfiguration().uiMode & Configuration.UI_MODE_NIGHT_MASK;
         if (dayNightMode == Configuration.UI_MODE_NIGHT_NO) {